fix(metadata): expose viteMetadata type
