+++ packages/vite/src/node/plugins/metadata.ts
@@ -1,10 +1,5 @@
 import type { Plugin } from '../plugin'
 
-export interface ChunkMetadata {
-  importedAssets: Set<string>
-  importedCss: Set<string>
-}
-
 /**
  * Prepares the rendered chunks to contain additional metadata during build.
  */
